requirements.txt
PyQt5
numpy
scipy
Pillow
pyyaml
